refactor(minesweeper): MineSweeperGame 에서 ui 입출력 분리
- ConsoleInput: 입력 담당 UI
- ConsoleOutput: 출력 담당 UI
